Pipeline generation as plugins (#321)
