+++ .github/pull_request_template.md
@@ -1,0 +1,15 @@
+## Overview
+Author: 
+
+## Issue 
+Issue link (if present):
+Issue description:
+Related PRs
+
+## Changes
+ROS distro:
+List of changes:
+
+## Testing
+Hardware used:
+Depthai library version:
