+++ depthai-ros/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.10.2)  # CMake version in Ubuntu 18.04 LTS
 
-project(depthai-ros VERSION 2.7.2 LANGUAGES CXX C)
+project(depthai-ros VERSION 2.7.3 LANGUAGES CXX C)
 
 set(CMAKE_CXX_STANDARD 14)
