+++ depthai_bridge/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package format="3">
   <name>depthai_bridge</name>
-  <version>2.7.2</version>
+  <version>2.7.3</version>
   <description>The depthai_bridge package</description>
 
   <maintainer email="sachin@luxonis.com">Sachin Guruswamy</maintainer>
