+++ depthai_descriptions/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package format="3">
   <name>depthai_descriptions</name>
-  <version>2.7.2</version>
+  <version>2.7.3</version>
   <description>The depthai_descriptions package</description>
 
   <maintainer email="sachin@luxonis.com">Sachin Guruswamy</maintainer>
