+++ depthai_ros_driver/config/camera.yaml
@@ -18,6 +18,7 @@
       i_acc_freq: 400
       i_acc_cov: 0.0
       i_batch_report_threshold: 1
+      i_get_base_device_timestamp: false
       i_enable_rotation: false
       i_gyro_cov: 0.0
       i_gyro_freq: 400
@@ -31,7 +32,7 @@
     left:
       i_board_socket_id: 1
       i_calibration_file: ''
-      i_get_base_device_timestamp: true
+      i_get_base_device_timestamp: false
       i_fps: 30.0
       i_height: 720
       i_low_bandwidth: false
@@ -53,7 +54,7 @@
     rgb:
       i_board_socket_id: 0
       i_simulate_from_topic: false
-      i_get_base_device_timestamp: true
+      i_get_base_device_timestamp: false
       i_disable_node: false
       i_calibration_file: ''
       i_enable_preview: false
@@ -82,7 +83,7 @@
     right:
       i_board_socket_id: 2
       i_calibration_file: ''
-      i_get_base_device_timestamp: true
+      i_get_base_device_timestamp: false
       i_fps: 30.0
       i_height: 720
       i_low_bandwidth: false
@@ -98,7 +99,7 @@
       r_set_man_exposure: false
     stereo:
       i_align_depth: true
-      i_get_base_device_timestamp: true
+      i_get_base_device_timestamp: false
       i_output_disparity: false
       i_bilateral_sigma: 0
       i_board_socket_id: 0
