+++ depthai_ros_driver/include/depthai_ros_driver/dai_nodes/nn/detection.hpp
@@ -43,9 +43,19 @@ class Detection : public BaseNode {
     void setupQueues(std::shared_ptr<dai::Device> device) override {
         nnQ = device->getOutputQueue(nnQName, ph->getParam<int>("i_max_q_size"), false);
         auto tfPrefix = getTFPrefix("rgb");
+        int width;
+        int height;
+        if(ph->getParam<bool>("i_disable_resize")) {
+            width = getROSNode()->get_parameter("rgb.i_preview_size").as_int();
+            height = getROSNode()->get_parameter("rgb.i_preview_size").as_int();
+        }
+        else{
+            width = imageManip->initialConfig.getResizeConfig().width;
+            height = imageManip->initialConfig.getResizeConfig().height;
+        }
         detConverter = std::make_unique<dai::ros::ImgDetectionConverter>(tfPrefix + "_camera_optical_frame",
-                                                                         imageManip->initialConfig.getResizeConfig().width,
-                                                                         imageManip->initialConfig.getResizeConfig().height,
+                                                                         width,
+                                                                         height,
                                                                          false,
                                                                          ph->getParam<bool>("i_get_base_device_timestamp"));
         detPub = getROSNode()->template create_publisher<vision_msgs::msg::Detection2DArray>("~/" + getName() + "/detections", 10);
@@ -59,9 +69,9 @@ class Detection : public BaseNode {
             infoManager->setCameraInfo(sensor_helpers::getCalibInfo(getROSNode()->get_logger(),
                                                                     *imageConverter,
                                                                     device,
-                                                                    dai::CameraBoardSocket::RGB,
-                                                                    imageManip->initialConfig.getResizeWidth(),
-                                                                    imageManip->initialConfig.getResizeWidth()));
+                                                                    static_cast<dai::CameraBoardSocket>(ph->getParam<int>("i_board_socket_id")),
+                                                                    width,
+                                                                    height));
 
             ptPub = image_transport::create_camera_publisher(getROSNode(), "~/" + getName() + "/passthrough/image_raw");
             ptQ->addCallback(std::bind(sensor_helpers::imgCB, std::placeholders::_1, std::placeholders::_2, *imageConverter, ptPub, infoManager));
