+++ depthai_ros_driver/src/camera.cpp
@@ -5,7 +5,7 @@
 #include "ament_index_cpp/get_package_share_directory.hpp"
 #include "depthai/device/Device.hpp"
 #include "depthai/pipeline/Pipeline.hpp"
-#include "depthai_ros_driver/pipeline_generator.hpp"
+#include "depthai_ros_driver/pipeline/pipeline_generator.hpp"
 
 namespace depthai_ros_driver {
