+++ depthai_ros_msgs/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package format="3">
   <name>depthai_ros_msgs</name>
-  <version>2.7.2</version>
+  <version>2.7.3</version>
   <description>Package to keep interface independent of the driver</description>
 
   <maintainer email="sachin@luxonis.com">Sachin Guruswamy</maintainer>
